fix: update date formatting to use short month in multiple components

+++ apps/web/src/features/education/EducationPage.tsx
@@ -53,13 +53,13 @@ export default function EducationPage() {
                     <time>
                       {new Date(item.startDate).toLocaleDateString('en-US', {
                           year: 'numeric',
-                          month: 'long',
+                          month: 'short',
                         })}{' '}
                       -{' '}
                       {item.endDate
                         ? new Date(item.endDate).toLocaleDateString('en-US', {
                             year: 'numeric',
-                            month: 'long',
+                            month: 'short',
                           })
                         : 'Present'}
                     </time>
@@ -154,7 +154,7 @@ export default function EducationPage() {
                       <time>
                         {new Date(item.date).toLocaleDateString('en-US', {
                           year: 'numeric',
-                          month: 'long',
+                          month: 'short',
                         })}
                       </time>
                       <span>•</span>
